split out events related to room subscriptions, activity, elastic, notifications etc to a separate service
