+++ src/common/enums/room.type.ts
@@ -1,7 +1,8 @@
 export enum RoomType {
-  DISCUSSION = 'discussion',
   POST = 'post',
   CALENDAR_EVENT = 'calendar_event',
-  FORUM_DISCUSSION = 'forum_discussion',
+  DISCUSSION = 'discussion',
+  DISCUSSION_FORUM = 'discussion_forum',
   UPDATES = 'updates',
+  CALLOUT = 'callout',
 }
